UPDATE NL 12/04/2025 16:00

+++ src/main/java/payPalManager/utility/PaypalManagement.java
@@ -37,7 +37,7 @@ public final class PaypalManagement{
         return creationAPIToken.execute();
     }
 
-    public static Optional<PaypalOrdersCreated> createOrder(String accessToken, long userId, String apiBaseURL, String currency, double amount, String locale, String returnUrl, String cancelUrl){
+    public static Optional<PaypalOrdersCreated> createOrder(String accessToken, int userId, String apiBaseURL, String currency, double amount, String locale, String returnUrl, String cancelUrl){
         PaypalAPIRequest<PaypalOrdersCreated> creationOrder = new CreateOrderRequest.Builder()
                 .setApiBaseURL(apiBaseURL)
                 .setAccessToken(accessToken)
